DG-91 Ability to permanently delete schemas (#1389)
* DG-91 Ability to permamently delete schemas

* Removed unused enum

* Addressed PR comments

* Address PR comments

+++ core/src/main/java/io/confluent/kafka/schemaregistry/storage/KafkaSchemaRegistry.java
@@ -16,6 +16,20 @@
 package io.confluent.kafka.schemaregistry.storage;
 
 import io.confluent.kafka.schemaregistry.ParsedSchema;
+import io.confluent.kafka.schemaregistry.exceptions.IdDoesNotMatchException;
+import io.confluent.kafka.schemaregistry.exceptions.IncompatibleSchemaException;
+import io.confluent.kafka.schemaregistry.exceptions.InvalidSchemaException;
+import io.confluent.kafka.schemaregistry.exceptions.ReferenceExistsException;
+import io.confluent.kafka.schemaregistry.exceptions.SchemaRegistryException;
+import io.confluent.kafka.schemaregistry.exceptions.SchemaRegistryInitializationException;
+import io.confluent.kafka.schemaregistry.exceptions.SchemaRegistryTimeoutException;
+import io.confluent.kafka.schemaregistry.exceptions.SchemaVersionNotSoftDeletedException;
+import io.confluent.kafka.schemaregistry.exceptions.SubjectNotSoftDeletedException;
+import io.confluent.kafka.schemaregistry.exceptions.UnknownMasterException;
+import io.confluent.kafka.schemaregistry.exceptions.IdGenerationException;
+import io.confluent.kafka.schemaregistry.exceptions.SchemaRegistryStoreException;
+import io.confluent.kafka.schemaregistry.exceptions.OperationNotPermittedException;
+import io.confluent.kafka.schemaregistry.exceptions.SchemaRegistryRequestForwardingException;
 import org.apache.avro.reflect.Nullable;
 import org.apache.kafka.clients.admin.AdminClient;
 import org.apache.kafka.clients.admin.AdminClientConfig;
@@ -63,18 +77,6 @@ import io.confluent.kafka.schemaregistry.client.rest.entities.requests.RegisterS
 import io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientException;
 import io.confluent.kafka.schemaregistry.client.rest.utils.UrlList;
 import io.confluent.kafka.schemaregistry.client.security.SslFactory;
-import io.confluent.kafka.schemaregistry.exceptions.IdDoesNotMatchException;
-import io.confluent.kafka.schemaregistry.exceptions.IdGenerationException;
-import io.confluent.kafka.schemaregistry.exceptions.IncompatibleSchemaException;
-import io.confluent.kafka.schemaregistry.exceptions.InvalidSchemaException;
-import io.confluent.kafka.schemaregistry.exceptions.OperationNotPermittedException;
-import io.confluent.kafka.schemaregistry.exceptions.ReferenceExistsException;
-import io.confluent.kafka.schemaregistry.exceptions.SchemaRegistryException;
-import io.confluent.kafka.schemaregistry.exceptions.SchemaRegistryInitializationException;
-import io.confluent.kafka.schemaregistry.exceptions.SchemaRegistryRequestForwardingException;
-import io.confluent.kafka.schemaregistry.exceptions.SchemaRegistryStoreException;
-import io.confluent.kafka.schemaregistry.exceptions.SchemaRegistryTimeoutException;
-import io.confluent.kafka.schemaregistry.exceptions.UnknownMasterException;
 import io.confluent.kafka.schemaregistry.id.IdGenerator;
 import io.confluent.kafka.schemaregistry.id.IncrementalIdGenerator;
 import io.confluent.kafka.schemaregistry.id.ZookeeperIdGenerator;
@@ -565,7 +567,8 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
 
   @Override
   public void deleteSchemaVersion(String subject,
-                                  Schema schema)
+                                  Schema schema,
+                                  boolean permanentDelete)
       throws SchemaRegistryException {
     try {
       if (getModeInScope(subject) == Mode.READONLY) {
@@ -575,18 +578,27 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
       if (!lookupCache.referencesSchema(key).isEmpty()) {
         throw new ReferenceExistsException(key.toString());
       }
+      SchemaValue schemaValue = (SchemaValue) lookupCache.get(key);
+      if (permanentDelete && !schemaValue.isDeleted()) {
+        throw new SchemaVersionNotSoftDeletedException(subject, schema.getVersion().toString());
+      }
       // Ensure cache is up-to-date before any potential writes
       kafkaStore.waitUntilKafkaReaderReachesLastOffset(subject, kafkaStoreTimeoutMs);
-      SchemaValue schemaValue = new SchemaValue(schema);
-      schemaValue.setDeleted(true);
-      kafkaStore.put(key, schemaValue);
-      if (!getAllVersions(subject, false).hasNext()) {
-        if (getMode(subject) != null) {
-          deleteMode(subject);
+      if (!permanentDelete) {
+        schemaValue = new SchemaValue(schema);
+        schemaValue.setDeleted(true);
+        kafkaStore.put(key, schemaValue);
+        if (!getAllVersions(subject, false).hasNext()) {
+          if (getMode(subject) != null) {
+            deleteMode(subject);
+          }
+          if (getCompatibilityLevel(subject) != null) {
+            deleteSubjectCompatibility(subject);
+          }
         }
-        if (getCompatibilityLevel(subject) != null) {
-          deleteSubjectCompatibility(subject);
-        }
+      } else {
+
+        kafkaStore.put(key, null);
       }
     } catch (StoreTimeoutException te) {
       throw new SchemaRegistryTimeoutException("Write to the Kafka store timed out while", te);
@@ -598,16 +610,17 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
 
   public void deleteSchemaVersionOrForward(
       Map<String, String> headerProperties, String subject,
-      Schema schema) throws SchemaRegistryException {
+      Schema schema, boolean permanentDelete) throws SchemaRegistryException {
 
     kafkaStore.lockFor(subject).lock();
     try {
       if (isMaster()) {
-        deleteSchemaVersion(subject, schema);
+        deleteSchemaVersion(subject, schema, permanentDelete);
       } else {
         // forward registering request to the master
         if (masterIdentity != null) {
-          forwardDeleteSchemaVersionRequestToMaster(headerProperties, subject, schema.getVersion());
+          forwardDeleteSchemaVersionRequestToMaster(headerProperties, subject,
+                  schema.getVersion(), permanentDelete);
         } else {
           throw new UnknownMasterException("Register schema request failed since master is "
                                            + "unknown");
@@ -619,7 +632,8 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
   }
 
   @Override
-  public List<Integer> deleteSubject(String subject) throws SchemaRegistryException {
+  public List<Integer> deleteSubject(String subject,
+                                     boolean permanentDelete) throws SchemaRegistryException {
     // Ensure cache is up-to-date before any potential writes
     try {
       if (getModeInScope(subject) == Mode.READONLY) {
@@ -628,23 +642,36 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
       kafkaStore.waitUntilKafkaReaderReachesLastOffset(subject, kafkaStoreTimeoutMs);
       List<Integer> deletedVersions = new ArrayList<>();
       int deleteWatermarkVersion = 0;
-      Iterator<Schema> schemasToBeDeleted = getAllVersions(subject, false);
+      Iterator<Schema> schemasToBeDeleted = getAllVersions(subject, permanentDelete);
       while (schemasToBeDeleted.hasNext()) {
         deleteWatermarkVersion = schemasToBeDeleted.next().getVersion();
         SchemaKey key = new SchemaKey(subject, deleteWatermarkVersion);
         if (!lookupCache.referencesSchema(key).isEmpty()) {
           throw new ReferenceExistsException(key.toString());
         }
+        if (permanentDelete) {
+          SchemaValue schemaValue = (SchemaValue) lookupCache.get(key);
+          if (!schemaValue.isDeleted()) {
+            throw new SubjectNotSoftDeletedException(subject);
+          }
+        }
         deletedVersions.add(deleteWatermarkVersion);
       }
-      DeleteSubjectKey key = new DeleteSubjectKey(subject);
-      DeleteSubjectValue value = new DeleteSubjectValue(subject, deleteWatermarkVersion);
-      kafkaStore.put(key, value);
-      if (getMode(subject) != null) {
-        deleteMode(subject);
-      }
-      if (getCompatibilityLevel(subject) != null) {
-        deleteSubjectCompatibility(subject);
+
+      if (!permanentDelete) {
+        DeleteSubjectKey key = new DeleteSubjectKey(subject);
+        DeleteSubjectValue value = new DeleteSubjectValue(subject, deleteWatermarkVersion);
+        kafkaStore.put(key, value);
+        if (getMode(subject) != null) {
+          deleteMode(subject);
+        }
+        if (getCompatibilityLevel(subject) != null) {
+          deleteSubjectCompatibility(subject);
+        }
+      } else {
+        for (Integer version : deletedVersions) {
+          kafkaStore.put(new SchemaKey(subject, version), null);
+        }
       }
       return deletedVersions;
 
@@ -658,15 +685,18 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
 
   public List<Integer> deleteSubjectOrForward(
       Map<String, String> requestProperties,
-      String subject) throws SchemaRegistryException {
+      String subject,
+      boolean permanentDelete) throws SchemaRegistryException {
     kafkaStore.lockFor(subject).lock();
     try {
       if (isMaster()) {
-        return deleteSubject(subject);
+        return deleteSubject(subject, permanentDelete);
       } else {
         // forward registering request to the master
         if (masterIdentity != null) {
-          return forwardDeleteSubjectRequestToMaster(requestProperties, subject);
+          return forwardDeleteSubjectRequestToMaster(requestProperties,
+                  subject,
+                  permanentDelete);
         } else {
           throw new UnknownMasterException("Register schema request failed since master is "
                                            + "unknown");
@@ -769,13 +799,15 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
   private void forwardDeleteSchemaVersionRequestToMaster(
       Map<String, String> headerProperties,
       String subject,
-      Integer version) throws SchemaRegistryRequestForwardingException {
+      Integer version,
+      boolean permanentDelete) throws SchemaRegistryRequestForwardingException {
     UrlList baseUrl = masterRestService.getBaseUrls();
 
     log.debug(String.format("Forwarding deleteSchemaVersion schema version request %s-%s to %s",
                             subject, version, baseUrl));
     try {
-      masterRestService.deleteSchemaVersion(headerProperties, subject, String.valueOf(version));
+      masterRestService.deleteSchemaVersion(headerProperties, subject,
+              String.valueOf(version), permanentDelete);
     } catch (IOException e) {
       throw new SchemaRegistryRequestForwardingException(
           String.format(
@@ -788,13 +820,14 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
 
   private List<Integer> forwardDeleteSubjectRequestToMaster(
       Map<String, String> requestProperties,
-      String subject) throws SchemaRegistryRequestForwardingException {
+      String subject,
+      boolean permanentDelete) throws SchemaRegistryRequestForwardingException {
     UrlList baseUrl = masterRestService.getBaseUrls();
 
     log.debug(String.format("Forwarding delete subject request for  %s to %s",
                             subject, baseUrl));
     try {
-      return masterRestService.deleteSubject(requestProperties, subject);
+      return masterRestService.deleteSubject(requestProperties, subject, permanentDelete);
     } catch (IOException e) {
       throw new SchemaRegistryRequestForwardingException(
           String.format(
@@ -877,13 +910,20 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
     final int version = versionId.getVersionId();
     Schema schema = this.get(subject, version, returnDeletedSchema);
     if (schema == null) {
-      if (!this.hasSubjects(subject)) {
+      if (!this.hasSubjects(subject, returnDeletedSchema)) {
         throw Errors.subjectNotFoundException(subject);
       } else {
         throw Errors.versionNotFoundException(version);
       }
     }
     return schema;
+  }
+
+  public boolean schemaVersionExists(String subject, VersionId versionId, boolean
+          returnDeletedSchema) throws SchemaRegistryException {
+    final int version = versionId.getVersionId();
+    Schema schema = this.get(subject, version, returnDeletedSchema);
+    return (schema != null);
   }
 
   @Override
@@ -971,10 +1011,12 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
   }
 
   @Override
-  public Set<String> listSubjects() throws SchemaRegistryException {
+  public Set<String> listSubjects(boolean returnDeletedSubjects)
+          throws SchemaRegistryException {
     try {
       Iterator<SchemaRegistryKey> allKeys = kafkaStore.getAllKeys();
-      return extractUniqueSubjects(allKeys);
+      return extractUniqueSubjects(allKeys,
+              returnDeletedSubjects);
     } catch (StoreException e) {
       throw new SchemaRegistryStoreException(
           "Error from the backend Kafka store", e);
@@ -1042,7 +1084,8 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
         .collect(Collectors.toList());
   }
 
-  private Set<String> extractUniqueSubjects(Iterator<SchemaRegistryKey> allKeys)
+  private Set<String> extractUniqueSubjects(Iterator<SchemaRegistryKey> allKeys,
+                                            boolean returnDeletedSubjects)
       throws StoreException {
     Set<String> subjects = new HashSet<String>();
     while (allKeys.hasNext()) {
@@ -1050,7 +1093,8 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
       if (k instanceof SchemaKey) {
         SchemaKey key = (SchemaKey) k;
         SchemaValue value = (SchemaValue) kafkaStore.get(key);
-        if (value != null && !value.isDeleted()) {
+        if (value != null
+                && (!value.isDeleted() || returnDeletedSubjects)) {
           subjects.add(key.getSubject());
         }
       }
@@ -1058,9 +1102,11 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
     return subjects;
   }
 
-  public boolean hasSubjects(String subject) throws SchemaRegistryStoreException {
+  public boolean hasSubjects(String subject,
+                             boolean lookupDeletedSubjects)
+          throws SchemaRegistryStoreException {
     try {
-      return lookupCache.hasSubjects(subject);
+      return lookupCache.hasSubjects(subject, lookupDeletedSubjects);
     } catch (StoreException e) {
       throw new SchemaRegistryStoreException(
           "Error from the backend Kafka store", e);
@@ -1253,7 +1299,7 @@ public class KafkaSchemaRegistry implements SchemaRegistry, MasterAwareSchemaReg
       kafkaStore.waitUntilKafkaReaderReachesLastOffset(subject, kafkaStoreTimeoutMs);
       if (mode == Mode.IMPORT && getMode(subject) != Mode.IMPORT) {
         // Changing to import mode requires that no schemas exist with matching subjects.
-        if (hasSubjects(subject)) {
+        if (hasSubjects(subject, false)) {
           throw new OperationNotPermittedException("Cannot import since found existing subjects");
         }
         // At this point no schemas should exist with matching subjects.
